searchbar and user login Info componentized
